# Optimize allocations in types derived from DescribedList