--- a/src/Framing/Accepted.cs
+++ b/src/Framing/Accepted.cs
@@ -32,6 +32,16 @@ namespace Amqp.Framing
         {
         }
 
+        internal override void WriteField(ByteBuffer buffer, int index)
+        {
+            Fx.Assert(false, "Invalid field index");
+        }
+
+        internal override void ReadField(ByteBuffer buffer, int index, byte formatCode)
+        {
+            Fx.Assert(false, "Invalid field index");
+        }
+
         /// <summary>
         /// Returns a string that represents the current accepted outcome.
         /// </summary>