… src/Framing/Begin.cs
… src/Framing/Codec.cs
… src/Framing/Detach.cs
… src/Framing/Dispose.cs
… src/Framing/End.cs
… src/Framing/Error.cs
… src/Framing/Flow.cs
… src/Framing/Header.cs
… src/Framing/Modified.cs
… src/Framing/Open.cs
… src/Framing/Properties.cs
… src/Framing/Received.cs
… src/Framing/Rejected.cs
… src/Framing/Released.cs
… src/Framing/Source.cs
… src/Framing/Target.cs
… src/Framing/Transfer.cs
… src/Sasl/SaslChallenge.cs
… src/Sasl/SaslInit.cs
… src/Sasl/SaslMechanisms.cs
… src/Sasl/SaslOutcome.cs
… src/Sasl/SaslResponse.cs
… src/Transactions/Coordinator.cs
… src/Transactions/Declare.cs
… src/Transactions/Declared.cs
… src/Transactions/Discharge.cs
… src/Transactions/TransactionalState.cs
… src/Types/DescribedList.cs
… src/Types/Encoder.cs
… src/Types/RestrictedDescribed.cs