Adresses #369 

Performance and allocations look much better now, see this benchmark that encodes and decodes a Header, the only allocations left are the actual Header instances:

```
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.16299.1268 (1709/FallCreatorsUpdate/Redstone3)
Intel Xeon CPU E5-1620 v2 3.70GHz, 1 CPU, 8 logical and 4 physical cores
Frequency=3613278 Hz, Resolution=276.7570 ns, Timer=TSC
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3416.0
  DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3416.0
```

|    Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| EncodeOld | 537.83 ns | 8.927 ns | 7.454 ns | 0.0410 |     - |     - |     216 B |
| EncodeNew |  93.70 ns | 1.622 ns | 1.355 ns | 0.0137 |     - |     - |      72 B |
| DecodeOld | 252.36 ns | 3.412 ns | 3.025 ns | 0.0653 |     - |     - |     344 B |
| DecodeNew |  58.01 ns | 1.083 ns | 1.013 ns | 0.0137 |     - |     - |      72 B |


One behavioural change is introduced with the optimization. Previously null items at the end of the Fields list where not encoded. If I understood the AMQP spec correctly this is allowed but not strictly necessary. It could be accomplished with this method too, but would complicate the encoding a lot, so I didn't do it. I don't think the difference will be noticeable, but it will mean that a few more null bytes are sent over the wire now.

I also found 3 other places where allocations can be avoided (see other changesets).

BTW: What is the maximum C# version that can be used in this project?