@xinchen10 2019-08-01
VS 2015 is required to build some netmf projects. So sadly we cannot use any language features that are not supported by VS 2015. NETMF SDK (maybe the nanoF as well?) does not support generics so we cannot use nullable.