@Havret 2019-08-02
@xinchen10 
> NETMF SDK (maybe the nanoF as well?) does not support generics so we cannot use nullable.

This looks like quite a limitation. Is this the reason you are maintaining two libraries alongside (this one and azure-amqp?