@szehetner 2019-08-02
I see two options:
* Implement custom non-generic nullable types (NullableBoolean, NullableULong,...) to be used for the backing fields.
* Have seperate code files for NETMF for these classes and keep the previous implementation for NETMF.

I would prefer the second one, but this could possibly mean more maintenance effort in the future.  If you don't want to do this, I can implement the first one.